corrected RGB type

+++ inc/visualization.hpp
@@ -6,7 +6,7 @@
 #include <math.h>
 #include <SFML/Graphics.hpp>
 
-void C_to_rgb(const double Temperature, int &r, int &g, int &b, const double Temperature_limit);
+void C_to_rgb(const double Temperature, uint8_t &r, uint8_t &g, uint8_t &b, const double Temperature_limit);
 
 void draw(const int rows, const int colums, double matrix[100][100],
           sf::RenderWindow& window,
